requirements.txt
python-dotenv

chainlit==1.3.0rc1
openai
websockets

langchain
langchain-community
langchain-openai
faiss-cpu
pypdf